+++ crucible-syntax/CHANGELOG.md
@@ -1,5 +1,6 @@
 # next
 
+* Add `seq-reverse` syntax for reversing symbolic sequences.
 * The `Lang.Crucible.Syntax.Overrides` module has been removed from this library,
   as the overrides it provided (`proveObligations` and `crucible-print-assumption-state`)
   were only used in the `crucible-cli` test suite. These overrides have been moved to