+++ crucible/CHANGELOG.md
@@ -1,5 +1,11 @@
 # next
 
+* Add `reverseSymSequence` to `Lang.Crucible.Simulator.SymSequence`
+* **BREAKING:** Add `SequenceReverse` constructor to `App` in
+  `Lang.Crucible.CFG.Expr` for reversing symbolic sequences.
+* **BREAKING:** Change the signature of `getRecordedTrace` in
+  `Lang.Crucible.Simulator.RecordAndReplay` to remove the `evalBool` parameter,
+  as the implementation no longer concretizes the trace.
 * Add `withStateBackend` and additional infrastructure for including stack
   traces in `SimError` and `IsSymBackend`
 * Fix `onlineProve` to properly negate goals before checking satisfiability